python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.7.4

# LangChain and AI - Compatible versions
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.15
langchain-cohere>=0.3.0
cohere==5.5.6
groq==0.4.2
langchain-groq==0.2.1

# Vector store
qdrant-client==1.7.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# HTTP client
httpx==0.25.2

# Utilities
python-multipart==0.0.6

# Web Search and HTTP
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Optional APIs for web search (set environment variables to enable)
# SERPAPI_KEY for Google Search
# BING_SEARCH_KEY for Bing Search 